+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/changefeed/common/ChangeFeedStartFromLegacyContinuationImpl.java
@@ -25,7 +25,7 @@ class ChangeFeedStartFromLegacyContinuationImpl extends ChangeFeedStartFromInter
     }
 
     @Override
-    public void populateRequest(RxDocumentServiceRequest request) {
+    public void populateRequest(RxDocumentServiceRequest request, ChangeFeedMode changeFeedMode) {
         // We don't need to set any headers to start from the beginning
     }