+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/changefeed/incremental/PartitionProcessorFactoryImpl.java
@@ -109,7 +109,7 @@ class PartitionProcessorFactoryImpl implements PartitionProcessorFactory {
                     this.changeFeedProcessorOptions),
                 null);
         } else {
-            state = lease.getContinuationState(this.collectionResourceId, feedRange);
+            state = lease.getIncrementalContinuationState(this.collectionResourceId, feedRange);
         }
 
         ProcessorSettings settings = new ProcessorSettings(state, this.collectionSelfLink)